requirements.txt
appnope==0.1.0
astroid==1.3.8
autohook==1.1.0
Babel==2.1.1
beautifulsoup4==4.4.1
behave==1.2.5
blinker==1.4
coverage==4.0.1
decorator==4.0.4
dominate==2.1.16
enum34==1.1.2
extras==0.0.3
Flask==0.10.1
Flask-Babel==0.9
Flask-Bootstrap==3.3.5.7
Flask-Login==0.3.2
Flask-Mail==0.9.1
flask-marshmallow==0.6.2
Flask-OAuth==0.12
Flask-OAuthlib==0.9.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.1
Flask-WhooshAlchemy==0.56
Flask-WTF==0.12
flipflop==1.0
funcsigs==0.4
fuzzywuzzy==0.7.0
gnureadline==6.3.3
guess-language==0.2
hgapi==1.7.2
httplib2==0.9.2
ipython==4.0.0
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
lettuce==0.2.21
linecache2==1.0.0
logilab-common==1.1.0
MarkupSafe==0.23
marshmallow==2.7.1
mock==1.3.0
nose==1.3.7
oauth2==1.9.0.post1
oauthlib==1.0.3
parse==1.6.6
parse-type==0.3.4
path.py==8.1.2
pbr==1.8.1
pexpect==4.0.1
pickleshare==0.5
ptyprocess==0.5
pylint==1.4.4
python-mimeparse==0.1.4
python-openid==2.2.5
python-subunit==1.2.0
pytz==2015.6
rauth==0.7.1
requests==2.8.1
requests-oauthlib==0.5.0
simplegeneric==0.8.1
six==1.10.0
speaklater==1.3
SQLAlchemy==1.0.9
sqlalchemy-migrate==0.10.0
sqlparse==0.1.17
sure==1.2.24
Tempita==0.5.2
testtools==1.8.0
traceback2==1.4.0
traitlets==4.0.0
unittest2==1.1.0
virtualenv==13.1.2
visitor==0.1.2
Werkzeug==0.10.4
Whoosh==2.7.0
WTForms==2.0.2
